Add javadoc copying mojo so we no longer need to generate javadoc into strange places.
Fix paths in lexakai mojo.

Lexakai mojo now strips SVG comments - otherwise it generates spurious, pointless diffs - it
includes the build date of PlantUML in a comment, but converts its time zone to local time,
so if Jon generates we get MST and if I do we get PST, which is a really silly reason to
create noise in your git history.