BumpVersionPlugin now works and produces results that are consistent and buildable.
One case still to deal with:  How to update versions of things that get changed as a side-effect:

 * Cactus is being updated to 1.9.9-SNAPSHOT
 * Since there is a cactus.version property in telenav-superpom, telenav-superpom gets a new version
 * Therefore, telenav-superpom gets its version bumped
 * Therefore, kivakit, mesakit, etc. superpoms get updated, so they need their version bumped
 * Therefore, every java project that uses one of _those_ superpoms needs an update

We handle this now, but we apply the version.flavor.change to _everything_ - meaning some things
become snapshots which probably shouldn't.  Need a test to see if a change is a side-effect or
explicitly requested, and do something slightly different for the ones that aren't.