Hang is caused by process failing to start (ulimit?  Seems a likely culprit).
Previous workaround was failing because we get an exit code back of -2147483648 and treated that
as programmer error in the caller, so an exception was thrown because of that before there was any
chance to notify the original caller.

Improved the retry loop.

Also switch the dev github action to jdk 17 to see if something JDK specific is involved.